[AMS-5] domain - 소속, 위치, 자산 유형 삭제 시 검증 추가 (현재 소프트 삭제이기 때문에 fk 제약 조건에 걸리지 않음 그래서 연관 관계 엔티티가 존재 할 경우 삭제 되지 않도록 함)

+++ domain/src/main/java/com/sewon/affiliation/application/AffiliationService.java
@@ -46,7 +46,8 @@ public class AffiliationService {
 
     @Transactional
     public void deleteAffiliationById(Long id) {
-        if (assetLocationRepository.findAllByAffiliationId(id).isEmpty()) {
+        if (assetLocationRepository.findAllByAffiliationId(id).isEmpty()
+            && !affiliationRepository.existsAccountByAffiliationId(id)) {
             affiliationRepository.deleteById(id);
             return;
         }
